refactor: move BcryptAdapter files to a folder
